add landing page service control for mmds
