+++ cdap-ui/app/cdap/components/Experiments/ListView/index.js
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2017-2018 Cask Data, Inc.
+ * Copyright © 2018 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -12,230 +12,86 @@
  * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
  * License for the specific language governing permissions and limitations under
  * the License.
-*/
+ */
 
-import PropTypes from 'prop-types';
-import React from 'react';
-import {connect} from 'react-redux';
-import LoadingSVGCentered from 'components/LoadingSVGCentered';
-import TopPanel from 'components/Experiments/TopPanel';
-import SortableStickyTable from 'components/SortableStickyTable';
-import PieChart from 'components/PieChart';
-import PaginationWithTitle from 'components/PaginationWithTitle';
-import d3 from 'd3';
-import ExperimentsListBarChart from 'components/Experiments/ListView/ExperimentsListBarChart';
-import PlusButton from 'components/PlusButton';
-import InvalidPageView from 'components/Experiments/ListView/InvalidPageView';
-import EmptyListView from 'components/Experiments/ListView/EmptyListView';
-import NamespaceStore, { getCurrentNamespace } from 'services/NamespaceStore';
-import { Link } from 'react-router-dom';
+import React, { Component } from 'react';
 import Helmet from 'react-helmet';
-import {handlePageChange} from 'components/Experiments/store/ActionCreator';
+import ExperimentsListViewWrapper from 'components/Experiments/ListView/ListViewWrapper';
+import { Provider } from 'react-redux';
+import experimentsStore, { DEFAULT_EXPERIMENTS } from 'components/Experiments/store';
+import { getExperimentsList, setAlgorithmsList, updatePagination, handlePageChange } from 'components/Experiments/store/ActionCreator';
+import queryString from 'query-string';
+import isNil from 'lodash/isNil';
+import Mousetrap from 'mousetrap';
 
-require('./ListView.scss');
-
-const tableHeaders = [
-  {
-    label: 'Experiment',
-    property: 'name'
-  },
-  {
-    label: '#Models',
-    property: 'numOfModels'
-  },
-  {
-    label: '#Deployed',
-    property: 'numOfDeployedModels'
-  },
-  {
-    label: 'Algorithm Types',
-    property: 'algorithmTypes'
-  },
-  {
-    label: 'Test Data',
-    property: 'testData'
+export default class ExperimentsList extends Component {
+  componentWillMount() {
+    setAlgorithmsList();
+    Mousetrap.bind('right', this.goToNextPage);
+    Mousetrap.bind('left', this.goToPreviousPage);
+    this.parseUrlAndUpdateStore();
+    getExperimentsList();
   }
-];
 
-const colorScale = d3.scale.category20();
-const PLUSBUTTONCONTEXTMENUITEMS = [
-  {
-    label: 'Create a new Experiment',
-    to: `/ns/${getCurrentNamespace()}/experiments/create`
+  componentWillUnmount() {
+    Mousetrap.unbind('left');
+    Mousetrap.unbind('right');
   }
-];
 
-const getAlgoDistribution = (models) => {
-  if (!models.length) {
-    return null;
+  componentWillReceiveProps(nextProps) {
+    this.parseUrlAndUpdateStore(nextProps);
+    getExperimentsList();
   }
-  let modelsMap = {};
-  models.forEach(model => {
-    let algo = model.algorithm;
-    if (!modelsMap[algo]) {
-      modelsMap = {
-        ...modelsMap,
-        [algo]: {
-          value: algo,
-          count: 1,
-          color: colorScale(algo)
-        }
-      };
-    } else {
-      modelsMap = {
-        ...modelsMap,
-        [algo]: {
-          ...modelsMap[algo],
-          count: modelsMap[algo].count + 1
-        }
-      };
+
+  goToNextPage = () => {
+    let {offset, limit, totalPages} = experimentsStore.getState().experiments;
+    let nextPage = offset === 0 ? 1 : Math.ceil((offset + 1) / limit);
+    if (nextPage < totalPages) {
+      handlePageChange({ selected: nextPage });
     }
-  });
-  return Object.keys(modelsMap).map(m => modelsMap[m]);
-};
-
-const renderTableBody = (entities) => {
-  let list = entities.map(entity => {
-    let models = entity.models || [];
-    let modelsCount = entity.modelsCount;
-    return {
-      name: entity.name,
-      description: entity.description,
-      numOfModels: modelsCount,
-      numOfDeployedModels: models.filter(model => model.deploytime).length,
-      testData: entity.srcpath.split('/').pop(),
-      algorithmTypes: getAlgoDistribution(models)
-    };
-  });
-  let {selectedNamespace: namespace} = NamespaceStore.getState();
-  return (
-    <table className="table">
-      <tbody>
-        {
-          list.map(entity => {
-            return (
-              <tr>
-                <Link to={`/ns/${namespace}/experiments/${entity.name}`}>
-                  <td>
-                    <h5>
-                      <div>{entity.name}</div>
-                      <small>{entity.description}</small>
-                    </h5>
-                  </td>
-                  <td>{entity.numOfModels}</td>
-                  <td>{entity.numOfDeployedModels}</td>
-                  <td>{!entity.algorithmTypes ? null : <PieChart data={entity.algorithmTypes} />}</td>
-                  <td>{entity.testData}</td>
-                </Link>
-              </tr>
-            );
-          })
-        }
-      </tbody>
-    </table>
-  );
-};
-
-const getDataForGroupedChart = (experiments) => {
-  if (!experiments.length) {
-    return null;
-  }
-  let data = [];
-  experiments.map(experiment => {
-    data.push(
-      {
-        name: experiment.name,
-        type: 'Models',
-        count: Array.isArray(experiment.models) ? experiment.models.length : 0
-      },
-      {
-        name: experiment.name,
-        type: 'Deployed',
-        count: Array.isArray(experiment.models) ? experiment.models.filter(model => model.deploytime).length : 0
-      }
-    );
-  });
-  return data;
-};
-
-function ExperimentsListView({ loading, list, totalPages, currentPage, totalCount }) {
-  if (loading) {
-    return <LoadingSVGCentered />;
-  }
-  let {selectedNamespace: namespace} = NamespaceStore.getState();
-  if (!list.length) {
-    return (
-      <div className="experiments-listview">
-        <TopPanel>
-          <h4>Analytics - All Experiments</h4>
-          <PlusButton
-            mode={PlusButton.MODE.resourcecenter}
-            contextItems={PLUSBUTTONCONTEXTMENUITEMS}
-          />
-        </TopPanel>
-        {
-          totalPages ? <InvalidPageView namespace={namespace} /> : <EmptyListView namespace={namespace} />
-        }
-
-      </div>
-    );
-  }
-  return (
-    <div className="experiments-listview">
-      <TopPanel>
-        <h4>Analytics - All Experiments</h4>
-        <PlusButton
-          mode={PlusButton.MODE.resourcecenter}
-          contextItems={PLUSBUTTONCONTEXTMENUITEMS}
-        />
-      </TopPanel>
-      <ExperimentsListBarChart
-        data={getDataForGroupedChart(list)}
-      />
-      <div className="clearfix">
-        <PaginationWithTitle
-          handlePageChange={handlePageChange}
-          currentPage={currentPage}
-          totalPages={totalPages}
-          title={totalCount > 1 ? "Experiments" : "Experiment"}
-          numberOfEntities={totalCount}
-        />
-        <SortableStickyTable
-          entities={list}
-          tableHeaders={tableHeaders}
-          renderTableBody={renderTableBody}
-        />
-      </div>
-    </div>
-  );
-}
-
-ExperimentsListView.propTypes = {
-  loading: PropTypes.bool,
-  list: PropTypes.arrayOf(PropTypes.object),
-  totalPages: PropTypes.number,
-  currentPage: PropTypes.number,
-  totalCount: PropTypes.number
-};
-
-const mapStateToProps = (state) => {
-  return {
-    loading: state.experiments.loading,
-    list: state.experiments.list,
-    totalPages: state.experiments.totalPages,
-    currentPage: state.experiments.offset === 0 ? 1 : Math.ceil((state.experiments.offset + 1) / state.experiments.limit),
-    totalCount: state.experiments.totalCount
   };
-};
 
-const ExperimentsListViewWrapper = connect(mapStateToProps)(ExperimentsListView);
+  goToPreviousPage = () => {
+    let {offset, limit} = experimentsStore.getState().experiments;
+    let prevPage = offset === 0 ? 1 : Math.ceil((offset + 1) / limit);
+    if (prevPage > 1) {
+      handlePageChange({ selected: prevPage - 2 });
+    }
+  };
 
-const ExperimentsWithTitle = () => (
-  <div>
-    <Helmet title="CDAP | All Experiments" />
-    <ExperimentsListViewWrapper />
-  </div>
-);
+  parseUrlAndUpdateStore = (nextProps) => {
+    let props = nextProps || this.props;
+    let { offset, limit } = this.getQueryObject(queryString.parse(props.location.search));
+    updatePagination({ offset, limit });
+  };
 
+  getQueryObject = (query) => {
+    if (isNil(query)) {
+      return {};
+    }
+    let {
+      offset = DEFAULT_EXPERIMENTS.offset,
+      limit = DEFAULT_EXPERIMENTS.limit
+    } = query;
+    offset = parseInt(offset, 10);
+    limit = parseInt(limit, 10);
+    if (isNaN(offset)) {
+      offset = DEFAULT_EXPERIMENTS.offset;
+    }
+    if (isNaN(limit)) {
+      limit = DEFAULT_EXPERIMENTS.limit;
+    }
+    return { offset, limit };
+  };
 
-export default ExperimentsWithTitle;
+  render() {
+    return (
+      <Provider store={experimentsStore}>
+        <div>
+          <Helmet title="CDAP | All Experiments" />
+          <ExperimentsListViewWrapper />
+        </div>
+      </Provider>
+    );
+  }
+}
