+++ cdap-ui/app/cdap/text/text-en.yaml
@@ -1088,6 +1088,24 @@ features:
     clearLabel: Clear
     title: 'No match found for "{searchText}"'
     suggestionTitle: "You can try to:"
+
+  Experiments:
+    ServiceControl:
+      Benefits:
+        b1: Intuitive Web UI for building, training, testing and evaluating Machine Learning models.
+        b2: Seamless, integrated experience from data preparation and cleansing to model building and deployment.
+        b3: Horizontal scalability over your Big Data environment.
+        b4: Out of the box support for common machine learning libraries like SparkML, DLF4J and H2O, along with support for deploying custom algorithms and libraries.
+        b5: Support for tuning custom hyperparameters for algorithms.
+        b6: Integrated metrics and visualization providing rich summaries and graphs for model evaluation.
+        title: "Some key benefits of MMDS are:"
+      checkMessage: Checking if MMDS is available...
+      contactMessage: "Contact support@cask.co to enable"
+      description: "Data Scientists typically build custom tooling for managing their machine learning models and deploying them. Model Management and Distribution Service (MMDS) provides a seamless, automated interface to help users develop, train, test, evaluate and deploy their machine learning models using CDAP."
+      enableBtnLabel: Enable MMDS
+      errorTitle: Enabling MMDS Failed
+      title: Welcome to Model Management and Distribution Service
+
   FileBrowser:
     directory: Directory
     EmptyMessage:
